Added ability to fit binned data (LM only)
It is now possible to supply a `bins` argument when creating an `Image` (or derived class). This is a segmentation map, with each integer value >=0 denoting the pixels that belong to a given bin. For example, one might have fluxes measured in Voronoi binned spectral data.

`Image` objects now have a `binned_data` attribute, which returns an array of the mean value within each bin. In the case of a `ModelImage`, the model is sampled on the usual pixel grid, before the mean is taken over each bin.

A similarly-binned weight or variance image can be supplied when creating a `TargetImage`.

Negative bin id values are ignored. A mask can be supplied as usual, in which case masked pixels are assigned a bin id of -1, and hence ignored.

The LM optimiser uses the `binned_data` attribute, instead of the `data` attribute, if the target image has a binning. This allows binned data to be fit in an appropriate manner.